Phylotrajectories

Documentation for Phylotrajectories.

Phylotrajectories.recombine

Phylotrajectories.sim_count_matrix

Phylotrajectories.tree_inference

Phylotrajectories.recombine — Method

function recombine(cluster_names::Vector{String}, count_matrix::Matrix{Int64}, dest::String, src::String)

Recombines cell clusters dest and src. Returns recombined cluster names and count matrix. Here recombination means taking a clonotype-wise sum of counts from count_matrix. src and its column in the count matrix are discarded in the recombined cluster names and count matrix respectively. dest *= "+" * src in the recombined cluster names.

Example

julia> cluster_names = ["Type$i" for i = 1:3];

julia> count_matrix = [i * j for i = 1:4, j = 1:3]
4×3 Matrix{Int64}:
 1  2   3
 2  4   6
 3  6   9
 4  8  12

julia> recombined_cluster_names, recombined_count_matrix = recombine(cluster_names, count_matrix, "Type1", "Type3");

julia> recombined_cluster_names
2-element Vector{String}:
 "Type1+Type3"
 "Type2"

julia> recombined_count_matrix
4×2 Matrix{Int64}:
  4  2
  8  4
 12  6
 16  8

Phylotrajectories.sim_count_matrix — Method

sim_count_matrix(
    tree::FelNode,
    n_clonotypes::Int,
    n_cells::Int,
    initial_partition::Partition,
    models;
    outpath = "",
)

Simulate a count matrix for a given tree. Returns the cluster names and a cell_types-by-clonotypes count matrix, where cell_types is the number of leaves on the tree and clonotypes <= n_clonotypes is the number of clonotypes whose sampled total count in the tree is non-zero. The count matrix sums to n_cells. models can either be a single BranchModel (if the model is global for all clonotypes) or an array of BranchModels, if the models are clonotype-wise, or a function that takes a clonotype index, and returns a FelNode -> Vector{<:BranchModel} function if you need clonotype-wise models to vary from one branch to another.

Phylotrajectories.tree_inference — Method

tree_inference(cluster_names::Vector{String}, cluster_clono_matrix::Matrix{Int64}; <keyword arguments>)

Find the Maximum Likelihood tree for a count matrix cluster_clono_matrix. Returns the ML tree, CTMC model, discretized states of frequencies, LL of the ML tree, and the final LLs of the initial trees (see n_random_trees).

Arguments

jump=1.0: the step size between frequency states (log domain).

a=1.0: the rate of transitions to lower frequencies.

b=1.0: the rate of transitions to higher frequencies.

Ne=1.0: the initial tree's effective population size.

sample_rate=10.0: the initial tree's sample rate.

start_branch_length=0.1: the initial tree's non-root branch lengths.

max_cycles=10: the number of topology-only optimization iterations.

n_random_trees=1: the number of initial tree samples.