Phylotrajectories

Documentation for Phylotrajectories.

Phylotrajectories.sim_count_matrix

Phylotrajectories.tree_inference

Phylotrajectories.sim_count_matrix — Method

sim_count_matrix(
    tree::FelNode,
    n_clonotypes::Int,
    n_cells::Int,
    initial_partition::Partition,
    models;
    outpath = "",
)

Simulate a count matrix for a given tree. Returns the cluster names and a cell_types-by-clonotypes count matrix, where cell_types is the number of leaves on the tree and clonotypes <= n_clonotypes is the number of clonotypes whose sampled total count in the tree is non-zero. The count matrix sums to n_cells. models can either be a single BranchModel (if the model is global for all clonotypes) or an array of BranchModels, if the models are clonotype-wise, or a function that takes a clonotype index, and returns a FelNode -> Vector{<:BranchModel} function if you need clonotype-wise models to vary from one branch to another.

Phylotrajectories.tree_inference — Method

tree_inference(cluster_names::Vector{String}, cluster_clono_matrix::Matrix{Int64}; <keyword arguments>)

Find the Maximum Likelihood tree for a count matrix cluster_clono_matrix. Returns the ML tree, CTMC model, the discretized states over frequencies, and the LL of the tree

Arguments

...