Phylotrajectories

Documentation for Phylotrajectories.

Phylotrajectories.ContinuousModel

Phylotrajectories.ContinuousUpdate

Phylotrajectories.DiscreteModel

Phylotrajectories.FrequencySampler

Phylotrajectories.GaussianStateSample

Phylotrajectories.OUGaussianStateSample

Phylotrajectories.RootAcceptanceRatio

Phylotrajectories.recombine

Phylotrajectories.sim_count_matrix

Phylotrajectories.tree_inference

Phylotrajectories.tree_inference

Phylotrajectories.ContinuousModel — Type

ContinuousModel(; <keyword arguments>)

Frequencies diffuse throughout the tree in a continuous space per Brownian motion. The posterior distribution over trees is explored with the Metropolis algorithm, see ContinuousUpdate for more details on this.

Keyword Arguments

update::ContinuousUpdate=ContinuousUpdate(): the characterization of an MCMC iteration.

mean_drift::Float64=0.0: mean drift parameter for the Brownian motion process in the unbounded frequency space. If update samples the mean drift, then this is the intitial value.

Ne::Float64=1.0: the initial tree's effective population size.

sample_rate::Float64=10.0: the initial tree's sample rate.

start_branch_length::Float64=0.1: the initial tree's non-root branch lengths.

n_samples::Int=10: number of MCMC samples to collect after burn-in.

burn_in::Int=1000: the number of MCMC iterations to discard as burn-in.

sample_interval::Int=10: the number of MCMC iterations between samples.

Phylotrajectories.ContinuousUpdate — Type

Summary

struct ContinuousUpdate <: MolecularEvolution.AbstractUpdate

Updates the leaf frequencies, phylogenetic tree, root state and position, and mean drift of the Brownian motion, with metropolis steps.

Constructor

ContinuousUpdate(; <keyword arguments>)

Keyword Arguments

branchlength_sampler::MolecularEvolution.BranchlengthSampler=Phylotrajectories.default_branchlength_sampler(): the proposal and prior distributions for branch length updates in MCMC.

frequency_sampler::FrequencySampler=FrequencySampler(Normal()): the proposal distribution for frequency updates in MCMC.

root_sampler::GaussianStateSample=GaussianStateSample(MvNormal(zeros(2), Diagonal([0.1, 0.1])), MvNormal(zeros(2), Diagonal([1.0, 0.1])), 1e-2, 1): the proposal and prior distributions for root updates in MCMC.

mean_drift_sampler::MeanDriftSampler=MeanDriftSampler(Normal(), Normal(-0.3, 0.5), 1.0): the proposal and prior distributions for mean drift updates in MCMC, and the variance drift of the Brownian motion.

models::Int=1: the number of consecutive models updates per MCMC iteration.

refresh::Bool=false: whether to refresh the messages in tree between update operations to ensure message consistency (for debugging purposes).

Note

To disable the sampling of the mean drift, one can set models=0.

Note

GaussianStateSample also updates the root position by default. See GaussianStateSample for more details.

Phylotrajectories.DiscreteModel — Type

DiscreteModel(; <keyword arguments>)

Frequencies are discretized into states, and evolve by a continuous time Markov chain (CTMC). You may choose between searching for the Maximum Likelihood tree, or sampling trees from the posterior distribution using the Metropolis algorithm.

Keyword Arguments

ML::Bool=true: whether to perform Maximum Likelihood inference or MCMC.

jump::Float64=CANONICAL_JUMP: the step size between frequency states (log domain).

a::Float64=1.0: the rate of transitions to lower frequencies.

b::Float64=1.0: the rate of transitions to higher frequencies.

Ne::Float64=1.0: the initial tree's effective population size.

sample_rate::Float64=10.0: the initial tree's sample rate.

start_branch_length::Float64=0.1: the initial tree's non-root branch lengths.

max_cycles::Int=10: the number of topology-only optimization iterations.

n_random_trees::Int=1: the number of initial tree n_samples (for Maximum Likelihood inference).

n_samples::Int=10: number of MCMC samples to collect.

burn_in::Int=1000: the number of MCMC iterations to discard as burn-in.

sample_interval::Int=10: the number of MCMC iterations between samples.

branchlength_sampler::MolecularEvolution.BranchlengthSampler=Phylotrajectories.default_branchlength_sampler(): the proposal and prior distributions for branch length updates in MCMC.

Phylotrajectories.FrequencySampler — Type

FrequencySampler(proposal::Distribution{Univariate,Continuous})

A type that allows you to specify an additive proposal function. It also holds the acceptance ratio accratio (accratio[1] stores the number of accepts, and acc_ratio[2] stores the number of rejects).

Phylotrajectories.GaussianStateSample — Type

Summary

struct GaussianStateSample <: MolecularEvolution.UniformRootPositionSample

Implements the metropolis algorithm for the global Gaussian parameters [μ, ν], where the root state of each clonotype is ~ N(μ, exp(ν)), and updates the root position by MolecularEvolution.UniformRootPositionSample if position = true. It also holds the acceptance ratio acc_ratio in an RootAcceptanceRatio struct.

Constructor

GaussianStateSample(proposal::ContinuousMultivariateDistribution, prior::ContinuousMultivariateDistribution, radius::Float64, consecutive::Int64; position::Bool = true)

Allows you to specify multivariate proposal and prior distributions for [μ, ν]. consecutive is the number of consecutive updates of the root (state and/or position) per MCMC iteration. radius (∈ [0,1]) scales tree's total branch length to determine local proposal radius for root position.

Phylotrajectories.OUGaussianStateSample — Type

Summary

struct OUGaussianStateSample <: MolecularEvolution.UniformRootPositionSample

Implements the metropolis algorithm for the global Gaussian parameters [μ, ν], where the root state of each clonotype is ~ N(μ, exp(ν)), and updates the root position by MolecularEvolution.UniformRootPositionSample if position = true. It also holds the acceptance ratio acc_ratio in an OURootAcceptanceRatio struct.

Constructor

OUGaussianStateSample(proposal::ContinuousMultivariateDistribution, prior::ContinuousMultivariateDistribution, radius::Float64, consecutive::Int64; position::Bool = true)

Allows you to specify multivariate proposal and prior distributions for [μ, ν]. consecutive is the number of consecutive updates of the root (state and/or position) per MCMC iteration. radius (∈ [0,1]) scales tree's total branch length to determine local proposal radius for root position.

Phylotrajectories.RootAcceptanceRatio — Type

mutable struct RootAcceptanceRatio

Holds tuples of (ratio, total, #acceptances) in the fields position and state, where ratio::Float64 is the acceptance ratio, total::Int64 is the total number of proposals, and #acceptances::Int64 is the number of acceptances.

Phylotrajectories.recombine — Method

function recombine(cluster_names::Vector{String}, count_matrix::Matrix{Int64}, dest::String, src::String)

Recombines cell clusters dest and src. Returns recombined cluster names and count matrix. Here recombination means taking a clonotype-wise sum of counts from count_matrix. src and its column in the count matrix are discarded in the recombined cluster names and count matrix respectively. dest *= "+" * src in the recombined cluster names.

Example

julia> cluster_names = ["Type$i" for i = 1:3];

julia> count_matrix = [i * j for i = 1:4, j = 1:3]
4×3 Matrix{Int64}:
 1  2   3
 2  4   6
 3  6   9
 4  8  12

julia> recombined_cluster_names, recombined_count_matrix = recombine(cluster_names, count_matrix, "Type1", "Type3");

julia> recombined_cluster_names
2-element Vector{String}:
 "Type1+Type3"
 "Type2"

julia> recombined_count_matrix
4×2 Matrix{Int64}:
  4  2
  8  4
 12  6
 16  8

Phylotrajectories.sim_count_matrix — Method

sim_count_matrix(
    tree::FelNode,
    n_clonotypes::Int,
    n_cells::Int,
    initial_partition::Partition,
    models;
    outpath = "",
)

Simulate a count matrix for a given tree. Returns the cluster names and a cell_types-by-clonotypes count matrix, where cell_types is the number of leaves on the tree and clonotypes <= n_clonotypes is the number of clonotypes whose sampled total count in the tree is non-zero. The count matrix sums to n_cells. models can either be a single BranchModel (if the model is global for all clonotypes) or an array of BranchModels, if the models are clonotype-wise, or a function that takes a clonotype index, and returns a FelNode -> Vector{<:BranchModel} function if you need clonotype-wise models to vary from one branch to another.

Phylotrajectories.tree_inference — Method

tree_inference(model::ContinuousModel, cluster_names::Vector{String}, cluster_clono_matrix::Matrix{Int64})

Returns the initial tree, sampled trees, LLs of the sampled trees, and the mean drift of the sampled models.

See ContinuousModel for model specification and parameters.

Phylotrajectories.tree_inference — Method

tree_inference(model::DiscreteModel, cluster_names::Vector{String}, cluster_clono_matrix::Matrix{Int64})

Find the Maximum Likelihood tree for a count matrix cluster_clono_matrix using a discrete state model. Returns the ML tree, CTMC model, discretized states of frequencies, optimized trees, and the final LLs of the initial trees.

See DiscreteModel for model specification and parameters.